ultimas alterações

+++ src/main/java/br/com/Fiap/filmeapi/model/Filme.java
@@ -1,12 +1,11 @@
 package br.com.Fiap.filmeapi.model;
 
-import java.math.BigDecimal;
 
 public record Filme(
 		Long id,
 		String titulo,
 		String sinopse,
 		String poster,
-		BigDecimal nota,
+		Double nota,
 		Boolean assistido
 ) {}
